+++ src/actions/wallet/signMessage.test.ts
@@ -89,7 +89,7 @@ test('error: no account', async () => {
     [AccountNotFoundError: Could not find an Account to execute with this Action.
     Please provide an Account with the \`account\` argument on the Action, or by supplying an \`account\` to the Client.
 
-    Docs: https://viem.sh/docs/actions/wallet/signMessage#account
+    Docs: https://viem.sh/docs/actions/wallet/signMessage
     Version: viem@x.y.z]
   `,
   )