+++ src/chains/definitions/abstractTestnet.ts
@@ -1,0 +1,23 @@
+import { defineChain } from '../../utils/chain/defineChain.js'
+import { chainConfig } from '../../zksync/chainConfig.js'
+
+export const abstractTestnet = /*#__PURE__*/ defineChain({
+  ...chainConfig,
+  id: 11_124,
+  name: 'Abstract Testnet',
+  nativeCurrency: {
+    decimals: 18,
+    name: 'ETH',
+    symbol: 'ETH',
+  },
+  rpcUrls: {
+    default: { http: ['https://api.testnet.abs.xyz'] },
+  },
+  blockExplorers: {
+    default: {
+      name: 'Abstract Block Explorer',
+      url: 'https://explorer.testnet.abs.xyz',
+    },
+  },
+  testnet: true,
+})