+++ src/clients/transports/webSocket.test.ts
@@ -185,6 +185,39 @@ test('subscribe', async () => {
   expect(blocks.length).toBe(1)
 })
 
+test('throws on socket closure', async () => {
+  const transport = webSocket(anvilMainnet.rpcUrl.ws, {
+    key: 'jsonRpc',
+    name: 'JSON RPC',
+  })({})
+  if (!transport.value) return
+
+  let error: Error | undefined
+  const { subscriptionId } = await transport.value.subscribe({
+    params: ['newHeads'],
+    onData: () => {},
+    onError: (error_) => {
+      error = error_
+    },
+  })
+
+  // Make sure we are subscribed.
+  expect(subscriptionId).toBeDefined()
+
+  await wait(100)
+  const { socket } = await transport.value.getRpcClient()
+  socket.close()
+  await wait(100)
+
+  expect(error).toMatchInlineSnapshot(`
+    [SocketClosedError: The socket has been closed.
+
+    URL: http://localhost
+
+    Version: viem@x.y.z]
+  `)
+})
+
 test('throws on bogus subscription', async () => {
   const transport = webSocket(anvilMainnet.rpcUrl.ws, {
     key: 'jsonRpc',
@@ -205,7 +238,7 @@ test('throws on bogus subscription', async () => {
 
 test('no url', () => {
   expect(() => webSocket()({})).toThrowErrorMatchingInlineSnapshot(`
-    [ViemError: No URL was provided to the Transport. Please provide a valid RPC URL to the Transport.
+    [UrlRequiredError: No URL was provided to the Transport. Please provide a valid RPC URL to the Transport.
 
     Docs: https://viem.sh/docs/clients/intro
     Version: viem@x.y.z]