+++ src/errors/transport.ts
@@ -9,6 +9,7 @@ export class UrlRequiredError extends BaseError {
       'No URL was provided to the Transport. Please provide a valid RPC URL to the Transport.',
       {
         docsPath: '/docs/clients/intro',
+        name: 'UrlRequiredError',
       },
     )
   }